requirements.txt
flask
flask-cors
opencv-python
python_dotenv
ultralytics
vietocr